requirements.txt
# Flask and extensions
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-SocketIO==5.1.1
Flask-Migrate==3.1.0
python-socketio==5.5.2

# Database
SQLAlchemy==1.4.41

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.3
joblib>=1.1.0

# Network packet capture and analysis
scapy==2.5.0
netifaces==0.11.0

# Utilities
python-dotenv==0.19.2
PyYAML==6.0.1

# Development and testing
pytest==7.1.3
eventlet==0.33.3
